Refactor setUpActions (#110)
* Factored out setUpActions helpers and sorted actions by task lifecycle
